requirements.txt
requests
pytest
jsonpath-ng
xmltodict
lxml
jsonschema
requests-oauthlib
openpyxl
faker
allure-pytest
pytest-html
pytest-xdist
pytest-dependency
pytest-rerunfailures
playwright
pytest
pytest-playwright
allure-pytest